// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  PROVIDER: 'PROVIDER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const RoomType = {
  MEETING_ROOM: 'MEETING_ROOM',
  PRIVATE_OFFICE: 'PRIVATE_OFFICE',
  PODCAST_STUDIO: 'PODCAST_STUDIO'
} as const

export type RoomType = (typeof RoomType)[keyof typeof RoomType]


export const BookingStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const BlockReason = {
  MAINTENANCE: 'MAINTENANCE',
  HOLIDAY: 'HOLIDAY'
} as const

export type BlockReason = (typeof BlockReason)[keyof typeof BlockReason]